Stop importing requests-heavy mocks from the pandas Foldseek extractor
The pandas conda env has no requests, so mocked search-mode CI died in extract_foldseek_hits after it started importing tests.mocks.
